New tag argument to control prefix of version
Add an new input `tag` and remove `version` as required option to allow
smooth transition to the new `tag` argument later. If tag is set version
argument is ignored, if version argument is set a warning is emitted to
transition to tag and tag is set to its value.

Internally tag is added along version to differentiate the version
(strict semver) and the tag (potentially 'v' prefixed semver). If the
tag is prefixed, version is set as the non-prefixed tag value.